backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.4.0
python-multipart==0.0.9
pandas==2.2.2
great-expectations==0.18.14
celery==5.4.0
redis==5.0.8
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
